requirements.txt
numpy==1.26.0
scikit-learn==1.5.1
scikit-survival = 0.23.0
monai==1.3.2
lifelines == 0.29.0
matplotlib == 3.9.1
notebook==7.0.5
torchtuples==0.2.2
